+++ lib/p2p/Peer.ts
@@ -120,9 +120,9 @@ class Peer extends EventEmitter {
   /** Connection retries min delay. */
   private static readonly CONNECTION_RETRIES_MIN_DELAY = 5000;
   /** Connection retries max delay. */
-  private static readonly CONNECTION_RETRIES_MAX_DELAY = 300000;
+  private static readonly CONNECTION_RETRIES_MAX_DELAY = 3600000; // 1 hour
   /** Connection retries max period. */
-  private static readonly CONNECTION_RETRIES_MAX_PERIOD = 604800000;
+  private static readonly CONNECTION_RETRIES_MAX_PERIOD = 259200000; // 3 days
 
   /** The version of xud this peer is using, or an empty string if it is still not known. */
   public get version() {